moving newton test to integrations
